APPSRE-6609 reduce generated classes
