+++ reconcile/gql_definitions/fragments/resource_values.py
@@ -1,0 +1,26 @@
+"""
+Generated by qenerate plugin=pydantic_v1. DO NOT MODIFY MANUALLY!
+"""
+from collections.abc import Callable  # noqa: F401 # pylint: disable=W0611
+from enum import Enum  # noqa: F401 # pylint: disable=W0611
+from typing import (  # noqa: F401 # pylint: disable=W0611
+    Any,
+    Optional,
+    Union,
+)
+
+from pydantic import (  # noqa: F401 # pylint: disable=W0611
+    BaseModel,
+    Extra,
+    Field,
+    Json,
+)
+
+
+class ResourceValues(BaseModel):
+    cpu: Optional[str] = Field(..., alias="cpu")
+    memory: Optional[str] = Field(..., alias="memory")
+
+    class Config:
+        smart_union = True
+        extra = Extra.forbid
